requirements.txt
streamlit
google-generativeai
pandas>=2.0.0
numpy>=1.24.0
python-dotenv==1.0.1
pymupdf>=1.23.0
PyPDF2>=3.0.0
pillow
tqdm
requests>=2.31.0
markdown>=3.6
pytest
matplotlib

# LangGraph and LangChain - Compatible versions with enhanced features
langgraph==0.2.50
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.5
langchain-google-genai==2.0.5

# Additional LangChain utilities for output parsing and callbacks
pydantic-settings>=2.0.0  # For enhanced Pydantic integration
tiktoken>=0.5.0  # Token counting for cost estimation

# ML/Data Science dependencies
scikit-learn>=1.3.0
scipy>=1.11.0
xgboost

# Required dependencies
pydantic>=2.5.0
typing-extensions>=4.8.0
aiohttp>=3.9.0